+++ src/components/templates/Confirmation/index.tsx
@@ -1,0 +1,5 @@
+import ConfirmationMain from '../../organisms/ConfirmationMain';
+
+export default function Confirmation() {
+  return <ConfirmationMain />;
+}
